+++ src/scrapers/movie/AEBN.cpp
@@ -131,14 +131,15 @@ void AEBN::onSearchFinished()
     auto* reply = dynamic_cast<QNetworkReply*>(QObject::sender());
     if (reply == nullptr) {
         qCritical() << "[AEBN] onSearchFinished: nullptr reply | Please report this issue!";
-        emit searchDone({}, {ScraperError::Type::InternalError, tr("Internal Error: Please report!")});
+        emit searchDone(
+            {}, {ScraperError::Type::InternalError, tr("Internal Error: Please report!"), "nullptr dereference"});
         return;
     }
     reply->deleteLater();
 
     if (reply->error() != QNetworkReply::NoError) {
         qWarning() << "Network Error" << reply->errorString();
-        emit searchDone({}, {ScraperError::Type::NetworkError, reply->errorString()});
+        emit searchDone({}, mediaelch::replyToScraperError(*reply));
         return;
     }