+++ src/scrapers/movie/IMDB.cpp
@@ -138,14 +138,15 @@ void IMDB::onSearchFinished()
     auto* reply = dynamic_cast<QNetworkReply*>(QObject::sender());
     if (reply == nullptr) {
         qCritical() << "[IMDb] onSearchFinished: nullptr reply | Please report this issue!";
-        emit searchDone({}, {ScraperError::Type::InternalError, tr("Internal Error: Please report!")});
+        emit searchDone(
+            {}, {ScraperError::Type::InternalError, tr("Internal Error: Please report!"), "nullptr dereference"});
         return;
     }
     reply->deleteLater();
 
     if (reply->error() != QNetworkReply::NoError) {
         qWarning() << "[IMDb] Search: Network Error" << reply->errorString();
-        emit searchDone({}, {ScraperError::Type::NetworkError, reply->errorString()});
+        emit searchDone({}, mediaelch::replyToScraperError(*reply));
         return;
     }